projects/01_chat_with_pdf/requirements.txt
langchain
langgraph
openai
streamlit
chromadb
python-dotenv
pypdf
tiktoken